+++ src/Network/Network/Models/AzureFirewall/PSAzureFirewall.cs
@@ -96,6 +96,12 @@ namespace Microsoft.Azure.Commands.Network.Models
         public PSExtendedLocation ExtendedLocation { get; set; }
 
         [JsonIgnore]
+        public string ExtendedLocationText
+        {
+            get { return JsonConvert.SerializeObject(ExtendedLocation, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
         public string IpConfigurationsText
         {
             get { return JsonConvert.SerializeObject(IpConfigurations, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }