+++ src/Network/Network/ChangeLog.md
@@ -21,8 +21,8 @@
 ## Upcoming Release
 
 * Added EdgeZone (Extended Location) parameter support for Azure Firewall
-  * Added -EdgeZone parameter to New-AzFirewall cmdlet
-  * When EdgeZone is specified, zones property is automatically set to null
+    - Added `-EdgeZone` parameter to `New-AzFirewall` cmdlet
+    - When `-EdgeZone` is specified, availability zones are not supported for Azure Firewall resources
 * Onboarded `Microsoft.DataReplication/replicationVaults` to Private Link Common Cmdlets
 
 ## Version 7.25.0